backend/freelancenowbackend/requirements.txt
asgiref==3.8.1
Django==5.1.1
djangorestframework==3.15.1
psycopg2-binary==2.9.7
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
